moved models and consts
